+++ src/Extensions/Syndication/UpdatePeriod.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Rss\Extensions\Syndication;
+
+enum UpdatePeriod: string {
+  case HOURLY = "hourly";
+  case DAILY = "daily";
+  case WEAKLY = "weakly";
+  case MONTHLY = "monthly";
+  case YEARLY = "yearly";
+}
+?>