+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.11.0-dev
 - BC break: moved RssResponse to namespace Nexendrie\Rss\Bridges\NetteApplication
 - raised minimal version of PHP to 8.1
 - BC break: callbacks for lastBuildDate and pubDate now have to return DateTime
+- BC break: removed Syndication::UPDATE_PERIOD_* constants, updatePeriod in Syndication now accepts enum Nexendrie\Rss\Extensions\Syndication\UpdatePeriod
 
 Version 0.10.0
 - raised minimal version of PHP to 7.4